Ajustes finales en produccion, pedidos y bodegas, ajustes de campos de texto y nombre de menus

+++ database/migrations/2019_06_19_033234_crear_tabla_producto.php
@@ -17,7 +17,7 @@ class CrearTablaProducto extends Migration
             $table->bigIncrements('id');
             $table->string('imagen')->unique();
             $table->string('nombre')->unique();
-            $table->integer('precio');
+            $table->decimal('precio', 13, 2);
 
             $table->unsignedBigInteger('material_id');
             $table->foreign('material_id')->references('id')->on('material');
